refactor(rds): decompose module_utils/rds.py into _rds/ private subpackage (#3033)
Cherry-pick of a07e0ff8e from main to stable-10.
Conflict in plugins/module_utils/rds.py resolved by taking the
incoming re-export shim (stable-10 used the deprecated
ansible.module_utils._text import; the _rds/ subpackage uses
ansible.module_utils.common.text.converters which is available
in all Ansible versions supported by stable-10 >=2.16.0).
