+++ tests/integration/targets/rds_instance_upgrade/defaults/main.yml
@@ -7,5 +7,5 @@ db_instance_class: db.t3.micro
 allocated_storage: 20
 
 # For mariadb tests
-mariadb_engine_version: 11.4.5
-mariadb_engine_version_2: 11.4.7
+mariadb_engine_version: "11.4.10"
+mariadb_engine_version_2: "11.4.12"
